+++ frontend/src/lib/components/LemonModal/LemonModal.stories.tsx
@@ -4,7 +4,7 @@ import { LemonModal, LemonModalProps } from './LemonModal'
 import { LemonButton } from '../LemonButton'
 
 export default {
-    title: 'Lemon UI/Lemon Modal',
+    title: 'Lemon UI/LemonModal',
     component: LemonModal,
 } as ComponentMeta<typeof LemonModal>
 
@@ -15,17 +15,45 @@ export const _LemonModal: ComponentStory<typeof LemonModal> = (props: LemonModal
             <LemonButton type="primary" onClick={() => setIsOpen(true)}>
                 Show control panel
             </LemonButton>
-            <LemonModal visible={isOpen} onCancel={() => setIsOpen(false)} {...props}>
-                <section>
-                    <h5>Adventure zone</h5>
-                    <LemonButton type="primary">Go into hyperspace</LemonButton>
-                </section>
-                <section>
-                    <h5>Danger zone</h5>
-                    <LemonButton type="secondary" status="danger">
-                        Initiate self-destruct sequence
-                    </LemonButton>
-                </section>
+            <LemonModal
+                {...props}
+                isOpen={isOpen}
+                onClose={() => setIsOpen(false)}
+                title="My Amazing Modal"
+                description="Helpful description content here"
+                footer={
+                    <>
+                        <div className="flex-1">
+                            <LemonButton type="secondary">Tertiary action</LemonButton>
+                        </div>
+                        <LemonButton type="secondary">Secondary</LemonButton>
+                        <LemonButton type="primary">Primary</LemonButton>
+                    </>
+                }
+            >
+                <p>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse
+                    cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
+                    culpa qui officia deserunt mollit anim id est laborum.
+                </p>
+
+                <p>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse
+                    cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
+                    culpa qui officia deserunt mollit anim id est laborum.
+                </p>
+
+                <p>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore
+                    et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+                    aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse
+                    cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
+                    culpa qui officia deserunt mollit anim id est laborum.
+                </p>
             </LemonModal>
         </>
     )
