--- frontend/src/lib/components/LemonModalV2/LemonModalV2.scss
@@ -1,90 +1,0 @@
-.LemonModalV2__overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-
-    transition: background-color 250ms ease-out, backdrop-filter 250ms ease-out;
-    z-index: var(--z-modal);
-
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    &.ReactModal__Overlay--after-open {
-        background-color: rgba(0, 0, 0, 0.2);
-        backdrop-filter: blur(5px);
-    }
-
-    &.ReactModal__Overlay--before-close {
-        background-color: rgba(0, 0, 0, 0);
-        backdrop-filter: blur(0px);
-    }
-}
-
-.LemonModalV2 {
-    position: relative;
-    display: inline-block;
-    max-width: 80%;
-    min-width: 200px;
-    width: 480px;
-    margin: 1rem auto;
-    border-radius: var(--radius);
-    background-color: #fff;
-    box-shadow: 0 16px 16px -16px rgba(0, 0, 0, 0.35);
-
-    transition: opacity 250ms ease-out, transform 250ms ease-out;
-    transform: scale(0.85) translateY(25%);
-    opacity: 0;
-    display: flex;
-    flex-direction: column;
-    overflow-y: hidden;
-    max-height: 90%;
-
-    &.ReactModal__Content--after-open {
-        transform: scale(1) translateY(0);
-        opacity: 1;
-    }
-
-    &.ReactModal__Content--before-close {
-        transform: scale(0.85);
-        opacity: 0;
-    }
-
-    .LemonModal__closebutton {
-        position: absolute;
-        top: 1rem;
-        right: 1rem;
-    }
-
-    .LemonModal__content {
-        padding: 1rem 1.5rem;
-        overflow-y: auto;
-    }
-}
-
-.LemonModal__header {
-    margin: 1.5rem;
-    margin-bottom: 0;
-    padding-bottom: 0.5rem;
-    border-bottom: 1px solid var(--border);
-
-    h3 {
-        margin-bottom: 0;
-    }
-}
-
-.LemonModal__footer {
-    display: flex;
-    justify-content: flex-end;
-    gap: 0.5rem;
-    border-top: 1px solid var(--border);
-    margin: 1.5rem;
-    margin-top: 0rem;
-    padding-top: 1rem;
-}
-
-.LemonModal__header + .LemonModal__footer {
-    border-top: none;
-}
