--- frontend/src/lib/components/LemonModalV2/index.ts
@@ -1,2 +1,0 @@
-export { LemonModalV2 } from './LemonModalV2'
-export type { LemonModalV2Props } from './LemonModalV2'
